requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
Markdown==3.10.2
python-dotenv==1.2.2
sqlparse==0.5.5
